pyproject.toml
[project]
name = "paper-rag"
version = "0.1.0"
description = "Production-grade local RAG for academic PDFs"
authors = [{name = "PaperRAG Team"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "docling",
    "sentence-transformers",
    "faiss-cpu",
    "torch",
    "numpy",
    "pydantic>=2.5,<3.0",
    "typer>=0.9",
    "rich>=13.0",
    "tqdm>=4.66",
    "python-dotenv>=1.0,<2.0",
    "psutil>=5.9",
]

[project.optional-dependencies]
llm = [
    "transformers>=4.38",
    "accelerate>=0.27",
    "openai>=1.0,<2.0",
]
test = [
    "pytest>=8.0",
]

[project.scripts]
paperrag = "paper_rag.cli:app"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

